refactor(module-registry): split OCI cache integrity helpers (#1107)
